Adds functions for balance updating
